requirements.txt
streamlit
tensorflow
pandas
openpyxl
Pillow
requests